fastapi_server/requirements.txt
fastapi
uvicorn
moondream
pillow
python-multipart
python-dotenv
transformers
torch
pdf2image